requirements.txt
google-generativeai>=0.3.0
reportlab>=4.0.0
PyPDF2>=3.0.0
pdfplumber>=0.11.0

